Projects/Youtube-Transcribe-Summarizer/requirements.txt
streamlit
python-dotenv
google-generativeai
pathlib
youtube-transcript-api